bundlr disconnect hook
